Split waste_dashboard.rs (847 lines) into subdirectory modules
Break the monolithic waste_dashboard.rs into three focused files:
- mod.rs (46 lines): module declarations, re-exports, visibility resource, plugin
- formatting.rs (101 lines): fmt_tons, fmt_pct, fmt_dollars helpers with tests
- warnings.rs (250 lines): WarningSeverity, severity functions, color mapping with tests
- dashboard_ui.rs (496 lines): main UI system, rendering helpers, UI-related tests

All pub items are re-exported from mod.rs so external API is unchanged.

Part of #1389
